fix color shade preview with dynamic routes
