+++ src/lib/types.ts
@@ -50,9 +50,3 @@ export type ColorAction = {
 };
 
 export type ColorState = ColorSpace & ColorAction;
-
-export type ColorShadeList = {
-  hex: HexColor;
-  rgb: RgbaColor;
-  label?: string;
-}[];
